requirements.txt
aliyun-python-sdk-core-v3==2.11.0
aliyun-python-sdk-dysmsapi==1.0.0
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1.1
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-import-export==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
future==0.16.0
httplib2==0.11.3
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
numpy==1.18.5
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.8
pandas==1.0.4
Pillow==5.3.0
psycopg2==2.7.5
pycrypto==2.6.1
PyJWT==1.6.4
PyMySQL==0.9.3
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
raven==6.9.0
requests==2.20.0
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.11.0
